Fix only_current_line with multiple buffers
Previously, whenever we handled showing diagnostics for a buffer, we
cleared our CursorMoved autocmd for all buffers, and thus showing the
current line diagnostic would stop no longer work in other buffers until
the LSP pushed the diagnostics for that buffer again (which would fix
that buffer but break all other again).

The fix is to only clear the autocmds for the buffer being handled.